requirements.txt
black==22.3.0
scikit-learn==1.0.2
streamlit==1.9.2
Pillow==9.1.1
numpy==1.22.1